pull request : backend 회의 내 모든 요구사항 반영

+++ naviSafe-backend/src/main/java/com/naviSafe/naviSafe/domain/outbreakOccur/dto/OutbreakResponseDto.java
@@ -14,8 +14,8 @@ public class OutbreakResponseDto {
     private ZonedDateTime occrDate;
     private ZonedDateTime expClrDate;
     private String accInfo;
-    private float grs80tmX;
-    private float grs80tmY;
+    private double grs80tmX;
+    private double grs80tmY;
 
     private String accTypeName;
     private String accDetailTypeName;
